refactor(engine): extract node-result/checkpoint/persistence to engine/persistence.rs (audit 🔴-1, PR4)
Fourth decomposition slice of the `engine.rs` god-module. The contiguous
cluster that turns a finished node into durable state moves verbatim into a
sibling `engine/persistence.rs`:
- `check_and_apply_idempotency` — idempotency short-circuit;
- `record_node_result` / `record_idempotency` — per-node result + key writes;
- `handle_panicked_node` — panic-to-failed-state handling;
- `checkpoint_node` / `checkpoint_node_port` — checkpoint writes (legacy +
  port seam);
- `persist_final_state` / `persist_final_state_port` — terminal-state
  persistence;
- `emit_final_event` — the terminal `ExecutionEvent`.

They stay `impl WorkflowEngine` methods in a child module (`use super::*`), so
no public API change. Eight methods called from sibling modules
(frontier/resume) or the parent (engine.rs, tests) are bumped private →
`pub(super)`; `checkpoint_node_port` stays private (only `checkpoint_node`
calls it, in-module).

`engine.rs`: 4,508 → 3,806 lines (13,298 → 3,806 cumulative across PR1–PR4,
-71%). Pure cut + the `mod persistence;` declaration + the visibility bumps.

Verified: `cargo fmt`/clippy `-p nebula-engine --all-features --all-targets`
clean; `cargo nextest run -p nebula-engine --all-features` → 442 passed,
0 skipped (identical set to pre-move).
